+++ src/_logger.ts
@@ -1,5 +1,5 @@
 import { Logger } from 'typescript-template-language-service-decorator'
-import * as ts from 'typescript/lib/tsserverlibrary'
+import type * as ts from 'typescript/lib/tsserverlibrary'
 
 import { pluginName } from './_config'