+++ src/index.ts
@@ -1,7 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import * as ts from 'typescript/lib/tsserverlibrary'
+import type * as ts from 'typescript/lib/tsserverlibrary'
 
 import { StyledPlugin } from './_plugin'
 
-export = (mod: { typescript: typeof ts }) => new StyledPlugin(mod.typescript)
+const init = (mod: { typescript: typeof ts }) => new StyledPlugin(mod.typescript)
+
+export { init as 'module.exports' }